# docs/swagger/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "RBAC scaffolding based on GIN + GORM + CASBIN + Dig.",
    "title": "GinAdmin",
    "contact": {
      "name": "LyricTian",
      "email": "tiannianshou@gmail.com"
    },
    "license": {},
    "version": "5.1.0"
  },
  "host": "127.0.0.1:10088",
  "basePath": "/",
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "查询数据",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer 用户令牌",
            "name": "Authorization",
            "in": "header"
          },
          {
            "type": "integer",
            "default": 1,
            "description": "分页索引",
            "name": "current",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "default": 10,
            "description": "分页大小",
            "name": "pageSize",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "角色名称(模糊查询)",
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询结果：{list:列表数据,pagination:{current:页索引,pageSize:页大小,total:总数量}}",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/schema.Role"
              }
            }
          },
          "401": {
            "description": "{error:{code:0,message:未授权}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "500": {
            "description": "{error:{code:0,message:服务器错误}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "创建数据",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer 用户令牌",
            "name": "Authorization",
            "in": "header"
          },
          {
            "description": "创建数据",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/schema.Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/schema.Role"
            }
          },
          "400": {
            "description": "{error:{code:0,message:无效的请求参数}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "401": {
            "description": "{error:{code:0,message:未授权}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "500": {
            "description": "{error:{code:0,message:服务器错误}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          }
        }
      }
    },
    "/api/v1/roles.select": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "查询选择数据",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer 用户令牌",
            "name": "Authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "查询结果：{list:角色列表}",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/schema.Role"
              }
            }
          },
          "400": {
            "description": "{error:{code:0,message:未知的查询类型}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "401": {
            "description": "{error:{code:0,message:未授权}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "500": {
            "description": "{error:{code:0,message:服务器错误}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          }
        }
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "查询指定数据",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer 用户令牌",
            "name": "Authorization",
            "in": "header"
          },
          {
            "type": "string",
            "description": "记录ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/schema.Role"
            }
          },
          "401": {
            "description": "{error:{code:0,message:未授权}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "404": {
            "description": "{error:{code:0,message:资源不存在}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "500": {
            "description": "{error:{code:0,message:服务器错误}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          }
        }
      },
      "put": {
        "tags": [
          "角色管理"
        ],
        "summary": "更新数据",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer 用户令牌",
            "name": "Authorization",
            "in": "header"
          },
          {
            "type": "string",
            "description": "记录ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "更新数据",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/schema.Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/schema.Role"
            }
          },
          "400": {
            "description": "{error:{code:0,message:无效的请求参数}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "401": {
            "description": "{error:{code:0,message:未授权}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "500": {
            "description": "{error:{code:0,message:服务器错误}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "角色管理"
        ],
        "summary": "删除数据",
        "parameters": [
          {
            "type": "string",
            "description": "Bearer 用户令牌",
            "name": "Authorization",
            "in": "header"
          },
          {
            "type": "string",
            "description": "记录ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{status:OK}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPStatus"
            }
          },
          "401": {
            "description": "{error:{code:0,message:未授权}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          },
          "500": {
            "description": "{error:{code:0,message:服务器错误}}",
            "schema": {
              "$ref": "#/definitions/schema.HTTPError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "schema.HTTPError": {
      "type": "object",
      "properties": {
        "error": {
          "description": "错误项",
          "type": "object",
          "$ref": "#/definitions/schema.HTTPErrorItem"
        }
      }
    },
    "schema.HTTPErrorItem": {
      "type": "object",
      "properties": {
        "code": {
          "description": "错误码",
          "type": "integer"
        },
        "message": {
          "description": "错误信息",
          "type": "string"
        }
      }
    },
    "schema.HTTPStatus": {
      "type": "object",
      "properties": {
        "status": {
          "description": "状态(OK)",
          "type": "string"
        }
      }
    },
...
    "schema.Role": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "created_at": {
          "description": "创建时间",
          "type": "string"
        },
        "creator": {
          "description": "创建者",
          "type": "string"
        },
        "memo": {
          "description": "备注",
          "type": "string"
        },
        "menus": {
          "description": "菜单权限",
          "type": "object",
          "$ref": "#/definitions/schema.RoleMenus"
        },
        "name": {
          "description": "角色名称",
          "type": "string"
        },
        "record_id": {
          "description": "记录ID",
          "type": "string"
        },
        "sequence": {
          "description": "排序值",
          "type": "integer"
        }
      }
    },
...
    "schema.RoleMenus": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "actions": {
            "description": "动作权限列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "menu_id": {
            "description": "菜单ID",
            "type": "string"
          },
          "resources": {
            "description": "资源权限列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}